moved handleTokenExchange to accessToken.tsx
